Refactor: 프로젝트 전체 타입스크립트 타입 추가
