+++ src/App.js
@@ -1,9 +1,10 @@
-import "./App.css";
+import React from 'react';
+import './App.css';
 
-import Main from "./pages/Main";
+import Main from './pages/Main.tsx';
 
-const App = () => {
+function App() {
   return <Main />;
-};
+}
 
 export default App;
