+++ src/components/todo/InputTodo.tsx
@@ -1,10 +1,12 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { FaPlusCircle, FaSpinner } from 'react-icons/fa';
 
-import { createTodo } from '../api/todo.tsx';
-import useFocus from '../hooks/useFocus.tsx';
+import { createTodo } from '../../api/todo.tsx';
+import useFocus from '../../hooks/useFocus.tsx';
 
-function InputTodo({ setTodos }) {
+import { Todo, SetTodos } from './TodoInterface.tsx';
+
+function InputTodo({ setTodos }: { setTodos: SetTodos }) {
   const [inputText, setInputText] = useState('');
   const [isLoading, setIsLoading] = useState(false);
   const { ref, setFocus } = useFocus();
@@ -28,7 +30,7 @@ function InputTodo({ setTodos }) {
         const { data } = await createTodo(newItem);
 
         if (data) {
-          return setTodos(prev => [...prev, data]);
+          return setTodos((prev: Todo[]) => [...prev, data]);
         }
       } catch (error) {
         console.error(error);
